fixed bugs. commented
NOTE: changed path for reading config file for project marking on
machines at school

+++ PathObject.java
@@ -1,23 +1,49 @@
 
+/**
+ * Class used as part of Distance Vector implementation. Representing the individual values (cells) in DV.
+ * @author James Alexander z3459695.
+ *
+ */
 public class PathObject {
 
 	private char viaNode;
-	private int distance;
+	private Double distance;
 	
-	public PathObject(char node, int d){
+	/**
+	 * Constructor.
+	 * @param node 	name of node corresponding to this entry in distance vector.
+	 * @param d		distance to node.
+	 */
+	public PathObject(char node, Double d){
 		viaNode = node;
 		distance = d;
 	}
+	/**
+	 * Getter method returning reference node corresponding to this entry in distance vector.
+	 * @return viaNode.
+	 */
 	public char getViaNode(){
 		return viaNode;
 	}
-	public int getDist(){
+	/**
+	 * Getter method returning reference distance to correspoding node (i.e represents specific value in DV).
+	 * @return distance.
+	 */
+	public Double getDist(){
 		return distance;
 	}
+	/**
+	 * Setter method to change value of node variable.
+	 * @param node	new name of node.
+	 */
 	public void setViaNode(char node){
 		viaNode = node;
 	}
-	public void setDist(int d){
+	/**
+	 * Setter method to change value of distance variable.
+	 * @param d	new distance to node.
+	 */
+	public void setDist(Double d){
 		distance = d;
 	}
 }
